LayoutTests/accessibility/ax-thread-text-apis/textarea-line-for-index.html

<!DOCTYPE HTML><!-- webkit-test-runner [ runSingly=true AccessibilityThreadTextApisEnabled=true ] -->
<!-- This is a modern-style rewrite of an existing test, and should replace it after AccessibilityThreadTextApisEnabled is enabled by default. -->
<html>
<head>
<script src="../../resources/accessibility-helper.js"></script>
<script src="../../resources/js-test.js"></script>
</head>
<body>

<textarea id="textarea" rows="5" cols="40">
line 1
line 2
line 3
</textarea>

<!-- Add another textarea so that we can ensure requesting lineForIndex from #textarea with an index outside its
     bounds doesn't spill into this textarea's lines. -->
<textarea rows="5" cols="40">
other text area
</textarea>

<script>
let output = "This test ensures basic LineForIndex functionality in textareas works correctly.\n\n";

if (window.accessibilityController) {
    const textarea = accessibilityController.accessibleElementById("textarea");
    for (let i = 0; i < 35; i++)
        output += `${i} -> ${textarea.lineForIndex(i)}\n`;

    debug(output);
}
</script>
</body>
</html>
