LayoutTests/accessibility/ax-thread-text-apis/textarea-line-for-index-expected.txt

This test ensures basic LineForIndex functionality in textareas works correctly.

0 -> 0
1 -> 0
2 -> 0
3 -> 0
4 -> 0
5 -> 0
6 -> 0
7 -> 1
8 -> 1
9 -> 1
10 -> 1
11 -> 1
12 -> 1
13 -> 1
14 -> 2
15 -> 2
16 -> 2
17 -> 2
18 -> 2
19 -> 2
20 -> 2
21 -> -1
22 -> -1
23 -> -1
24 -> -1
25 -> -1
26 -> -1
27 -> -1
28 -> -1
29 -> -1
30 -> -1
31 -> -1
32 -> -1
33 -> -1
34 -> -1

PASS successfullyParsed is true

TEST COMPLETE
